refactor: 테스트 코드 위치 변경
- ValidatorTest 제거
